requirements.txt
django
djangorestframework
pillow
pandas
scikit-learn  
matplotlib
seaborn
imblearn
statsmodels
keras
tensorflow